+++ src/code/matrix/evaluationfunctions/MatrixAStar1.java
@@ -6,7 +6,7 @@ import code.searchproblem.strategies.EvaluationFunction;
 public class MatrixAStar1 extends EvaluationFunction {
 
     @Override
-    public int evaluate(SearchTreeNode node) {
+    public long evaluate(SearchTreeNode node) {
         return node.pathCost + MatrixHeuristics.admissableHeuristic1(node);
     }
