+++ src/code/matrix/evaluationfunctions/MatrixAStar2.java
@@ -6,7 +6,7 @@ import code.searchproblem.strategies.EvaluationFunction;
 public class MatrixAStar2 extends EvaluationFunction {
 
     @Override
-    public int evaluate(SearchTreeNode node) {
+    public long evaluate(SearchTreeNode node) {
         return node.pathCost + MatrixHeuristics.admissableHeuristic2(node);
     }
